requirements.txt
chromadb>=0.3.28
sentence-transformers>=2.2.2
numpy>=1.23
scikit-learn>=1.2
python-dotenv>=1.0
tqdm>=4.64
nltk>=3.8
joblib>=1.2.0
